requirements.txt
# Dependencies for Sentiment Analysis Project
pandas==2.0.3
numpy==1.24.3
matplotlib==3.7.2
seaborn==0.12.2
flask==2.3.3
flask-cors==4.0.0
google-generativeai==0.3.1
python-dotenv==1.0.0
plotly==5.18.0
dash==2.14.1
dash-bootstrap-components==1.5.0
requests==2.31.0
beautifulsoup4==4.12.2
transformers==4.36.2
torch==2.1.2
scikit-learn==1.3.2